// src/app/models/Fournisseur/fournisseur.ts
export class Fournisseur {
  compte_tiers: string;
  intitule: string;
  adresse: string;
  complement: string;
  region: string;
  ville: string;
  pays: string;
  telephone: string;
  email: string;
  code_postal: string;
  ninea: string;

  constructor() {
    this.compte_tiers = "" ;
    this.intitule = "" ;
    this.adresse = "" ;
    this.complement = "" ;
    this.region = "" ;
    this.ville = "" ;
    this.pays = "" ;
    this.telephone = "" ;
    this.email = "" ;
    this.code_postal = "" ;
    this.ninea = "";

}
}